+++ src/ExtensionFactory.php
@@ -34,9 +34,10 @@ class ExtensionFactory implements ReferenceInterface
     const LATEST_PHP_5_5 = '5.5.38';
     const LATEST_PHP_5_6 = '5.6.40';
     const LATEST_PHP_7_0 = '7.0.33';
-    const LATEST_PHP_7_1 = '7.1.30';
-    const LATEST_PHP_7_2 = '7.2.20';
-    const LATEST_PHP_7_3 = '7.3.7';
+    const LATEST_PHP_7_1 = '7.1.33';
+    const LATEST_PHP_7_2 = '7.2.25';
+    const LATEST_PHP_7_3 = '7.3.12';
+    const LATEST_PHP_7_4 = '7.4.0';
 
     protected $storage;
 
@@ -182,7 +183,10 @@ class ExtensionFactory implements ReferenceInterface
         if (version_compare($phpVersion, '7.3', 'lt')) {
             return self::LATEST_PHP_7_2;
         }
-        return self::LATEST_PHP_7_3;
+        if (version_compare($phpVersion, '7.4', 'lt')) {
+            return self::LATEST_PHP_7_3;
+        }
+        return self::LATEST_PHP_7_4;
     }
 
     /**