+++ tests/Reference/Extension/CoreExtensionTest.php
@@ -41,7 +41,7 @@ class CoreExtensionTest extends GenericTest
      *
      * @return void
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         self::$optionalconstants = array(
             // Not real constant