+++ tests/Reference/Extension/LdapExtensionTest.php
@@ -41,7 +41,7 @@ class LdapExtensionTest extends GenericTest
      *
      * @return void
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         self::$optionalfunctions = array(
             // Requires LDAP SASL