+++ tests/Reference/Extension/McryptExtensionTest.php
@@ -43,7 +43,7 @@ class McryptExtensionTest extends GenericTest
      *
      * @return void
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         self::$optionalconstants = array(
             // Non-free