+++ tests/Reference/Extension/MemcacheExtensionTest.php
@@ -41,7 +41,7 @@ class MemcacheExtensionTest extends GenericTest
      *
      * @return void
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         self::$optionalconstants = array(
             'MEMCACHE_SERIALIZED'